Explicitly set gossip.endpoint in integration test
This change explicitly sets the advertised gossip endpoint to override
the address computed by the peer.

The configuration used for our integration test networks explicitly sets
the listen addresses to the IPv4 loopback address for endpoints other
than the chaincode server. The chaincode server listener is special
because it must be available to chaincode running in docker containers.

We configure the chaincode server to listen on INADDR_ANY and enable
address auto-detection so the peer can try to determine the address to
advertise for the chaincode server.

The auto-detect "feature" has a bizarre side effect in that it takes
precedence over an explicit configuration of the peer.address and this
impacts the advertised address for other subsystems like gossip.

Signed-off-by: Matthew Sykes <matthew.sykes@gmail.com>