Eliminate warnings when building documents
Prior to this patch, about 40 warnings occurs for some reasons when building
the documentations.
These warnings may be noisy and may prevent document maintainers or translators
from confirming for their new patches.

So, this patch fixes the documents to eliminate their warnings.

NOTE: Some "document isn't included in any toctree" errors still remain
because necessity of the target documents or the proper position of them in
the TOC is still unclear.

Signed-off-by: Tatsuya Sato <Tatsuya.Sato@hal.hitachi.com>