rm old proto related code from transient store
Before releasing v1.2, we changed the proto that we
used to store entries in the transient store.

1. v1.1 -- we stored only the private write-set
2. >=v1.2 -- we stored both private write-set and
   collection config

We had to retain the code related to both protos
because the transient store can have entries that
were stored during v1.1 (after it has been upgraded to v1.2).
This was done to ensure that we can do rolling upgrade
from v1.1 to v1.2 without any private data loss.

We don't need to maintain code that handles both protos.
Hence, this PR removes the code related to old proto.

FAB-18029

Signed-off-by: senthil <cendhu@gmail.com>