[FAB-16435] Improve gossip defaults - Peers to be leaders by default
Block dissemination via gossip may be removed in a future release,
since it is more straightforward for peers to simply pull blocks from ordering service.

To start the move in this direction, this PR changes the default for each
peer to be an org leader that pulls blocks from the ordering service, and
disables leader election by default.

Most integration tests will use the new defaults, including a lifecycle
integration test that adds an org3 to a channel at a later time, to ensure
that newly added orgs can receive blocks with the new defaults.

There is a gossip integration test that will continue to test both leader election and static leaders.
This PR updates the gossip integration test so that the new defaults are not used, to ensure
the non-default configurations continue to work.

Signed-off-by: David Enyeart <enyeart@us.ibm.com>