Restore couchdb logger
CouchDB client utility layer was recently refactored to be included in statecouchdb package.
This caused the traditional [couchdb] logger and [statecouchdb] logger
to log under a common [statecouchdb] logger.
Since administrators and support staff often want to see only the CouchDB interactions
in debug log, and since they are already accustomed to using the [couchdb] logger,
this change restores the separate [couchdb] logger for the CouchDB client code,
making it aligned with prior versions of Fabric.

Signed-off-by: David Enyeart <enyeart@us.ibm.com>