FAB-18065 Ch.Part.API: move follower to orderer/common
Following code review, it was decided to move the "follower" package 
from orderer/consensus/follower to orderer/common/follower.

The follower.Chain no longer implements consenus.Chain and is not
treated as a consensus plug-in.

Instead of a consensus.Consenter implemeting `JoinChain`, the registrar 
creates a follower.Chain and saves it in a separate map from
consensus.Chain implementations. Since the follower is not in the
chains map of the registrar, it is not found by broadcast requests
until it turns into an etcdraft.Chain.

Interface consensus.Consenter now defines a new interface
`IsChannelMember` that allows the registrar to detect whether a
join-block implies the orderer is a member or follower of the channel.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I12bd147961d70c6667aa93934ad9e936e2b4f03c