FAB-18086 Ch.Part.API: Improve wiring between Registrar and follower
Currently several opaque function pointers are used to decouple the
multichannel.Registrar and the follower.Chain. In this commit:

Convert opaque function pointers to well defined interfaces that are
easy to follow and mock for tests.

The BlockPuller creation mechanisms is converted to a factory struct
that receives all the prerequisites except the config-block, and exposes
a method that builds a new BlockPuller given a config block. This allows
us to use the same mechanism for on-boarding, until the join block, and
for following past the join-block. The factory also maintains a link to
a block signature verifier that is built from the latest config, and
exposes a method that allows it to update the block signature verifier
on every incoming config.

The block-puller was augmented with a method to update the endpoints.
The endpoints are updated on every incoming config past the join-block.

The follower's function pointer that was used to create a consensus.Chain
is replaced by an interface that defines this method.
This interface is implemented by the registrar and is set into the new
follower instance at construction time.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I3fa6e662c43ca5f566ae1644b8e61695da0ba2d4