Separate internal halt from Halt for etcdraft.Chain
(External) Halt stops a chain but does not call the haltCallback i
function. (Internal) halt stops a chain and calls the haltCallback,
which is used to transfer responsibility to a follower or inactive
chain. (External) Halt will be used by the channel participation API
when removing a channel.

FAB-18069 #done

Signed-off-by: Will Lahti <wtlahti@us.ibm.com>