Eventually before send metric checks in orderer
Sends through the comm layer, much like in gossip, are asynchronous.
Some of the metrics tests assumed that the completion of the "OnSubmit"
handler implied that the message send processing was complete while, in
reality, it only ensure the process was initiated. This left timing
windows where assertions related to send metrics would panic because
send processing had not completed and captured the related metrics.

Also, the test could not be focused because of the odd use of sub-tests
in `assertBiDiCommunicationForChannel` to handle setup of communication
streams that were associated with the wrong parent tests. This commit
undoes that craziness and wires the appropriate *testing.T to the
function to enable assertions at the correct scope.

Finally, to fix one of the issues with the metrics test, this uses a
`require.Eventually` to wait for the send metric to be available before
making assertions. It's likely there are other windows like this in the
tests but we'll leave those for another day.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>