Reuse PKCS11 objects for keys
Prior to this change, each invocation of Sign or
Verify actually made a PKCS11 call to find the
key object.  This is actually unnecessary as the
object handle is valid across multiple sessions.

This change adds a handle field to both the private
and public key objects and uses these handles across
multiple invocations of Sign and Verify for any given
key pair.

Signed-off-by: Gari Singh <gari.r.singh@gmail.com>