FAB-18122 BlockPuller goes into an endless loop on bad endpoint
If the `cluster.BlockPuller` receives a set of endpoint to which it
cannot connect (ever), and then `PullBlock()` is called, the calling
thread will go into an endless loop. The `tryFetchBlock` will try to
connect to some endpoint and will fails every time, sleep, and try
again without limit.

This prevents the follower.Chain (and possibly the etcdraft.Chain)
from implementing an exponential backoff and listening to the stop
signal. This means that this chain cannot be `Halt()` in order to
remove the follower.Chain when necessary, or stop it in tests.

It is better to return nil after a present number of retries rather
than get stuck in that loop for ever.

This commit includes a fix and a unit test that replicates the
problem and verifies it was fixed.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: Ic012bb02af75b2296c8b02a39179e9faf8ec7997