FAB-16233 more logs on flaky cluster test
Looking at the logs, one of the things I see is the message
"oops, something went wrong". This implies that the offending
test case is `TestBlockPullerFailures`. This is a table test
so it is hard to tell which test case is failing... however it
also says "Failed verifying received blocks: bad signature"
which means the failing test case is
"failure at verifying pulled block".

Lets add some additional logs around that test, and increase
its level to "debug" for the next time it fails.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: Ia38743ac926b4e4226995c2580ff9aed17bddeb2