Cache PKCS11 ObjectHandles
The pkcs11 implementation currently invokes
a find key operation for each sign and verify
operation.  This can be quite expensive especially
when communicating with network HSMs.

This change caches the ObjectHandle for private and
public keys as they should be valid across the current
PKCS11 session.

As ObjectHandles are only valid for the current session
context, if a stale session is found, the cache is cleared
as well.

Signed-off-by: Gari Singh <gari.r.singh@gmail.com>