FAB-18123 Ch.Part.API: Update BlockPuller endpoints (#1695)
As the follower pulls blocks beyond the join-block, it needs
to continuously apply the channel config from each config block
to the BlockPuller. One aspect of the channel config are the endpoints.

It is possible to simply recreate the block puller on every config
block. However this is inefficient because it drops the internal
buffer that is on average half-full of fetched and validated blocks.
In order to avoid this we add a method that updates the endpoints,
disconnects from the current connection (for safety), but does not
drop the internal buffer.

The next invocations of PullBlock will first re-connect, and then
pop blocks from the buffer until it is empty.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I68a6901b2d90ae276033de885954c6d3c6d445e7