Increase timeouts for discovery CLI and raft tests (#1758)
Locally, the discovery CLI test has been failing intermittently due to
an eventually timeout of 1s. While this should be enough, bumping it to
5 seconds seems to have resolved it.

In the raft migration tests, when the new orderer is started, the
network eventually timeout is not provided. This has resulted in a
couple of timeouts while waiting for the process to initialize.

Also, the grouper.Member used when invoking the orderer was unnecessary
so it's been removed.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>