FAB-18177 Flake in registrar unit test
Both the registrar_test.go and the block_puller.go factory do not assign value to

cluster.BlockPuller{
   MaxPullBlockRetries
   ...
}

which may cause an endless loop in cluster/deliver.go#L143.

Solution is to initialize a non zero value.

In addition, when a follower is started in registrar_test.go,
it must be stopped, otherwise the go-routine will run until  the process exits.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I4f20ebdaeb20e9dde24dc555c90e0684bfb5d720