Use proto message for statedb snapshot export (#1805)
In the current code, the internal serialized binary data for
the statedbs is exported to snapshot files. Though, this brings
performance efficiency, it compromises the code readability.
This also causes the binary data in snapshot files to differs for
leveldb and couchdb significantly. Moreover, while importing the
snapshot data during creation of a channel from snapshot, some of the
data needs to be interpreted outside the statedb packages.

This commit uses a common proto message for exporting the statedb data
for both leveldb and couchdb.

Signed-off-by: manish <manish.sethi@gmail.com>