Add logic for skipping checkouts for non-existent branches
Many users are confused by the checkout process for fabric-samples.
In particular when using the newest versions of fabric, there
is not always a corresponding fabric-samples tag to match
fabric. This adds logic to check for the existence of the tags
and lets the user know why we didn't checkout the tag

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>