Add object handle and key cache to bccsp/pkcs11
With this change, object handles obtained through findKeyPairFromSKI are
cached in the PKCS#11 implementation of BCCSP. Keys constructed by
getECKey are also cached.

In support of these changes, in addition to pooling idle sessions, the
provider tracks active sessions. If some condition occurs that results
in all sessions being closed, cached object handles are no longer valid
so the caches are purged.

Signed-off-by: Gari Singh <gari.r.singh@gmail.com>
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>