FAB-18204 Ch.Part.API: decouple multichannel and etcdraft
The etcdraft package imports the multichannel package.
This prevents the multichannel package from using the onboarding
and cluster packages, which depend on etcdraft.

Here etcdraft is decoupled from multichannel by declaring an
interface that capture the dependecy on the Registrar.

This also eliminates some of the opaque function pointers to the
Registrar that the etcdraft package was using.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: If8112edd7a8fe08a151bb3ca19bf0eeeedc2fb31