Add extra queryexecutor nil checks
Because system chaincode sometimes executes in a channel-less context,
there is a broken path throughout the chaincode package where despite
being passed as a parameter, the query executor is nil.

This path has been significantly cleaned up in the v2.0 code, though, so
long as system chaincode exists which attempts to execute without a
channel context, we will have this problem.

This small PR simply adds extra nil checks to prevent a crash in the
cases where a nil query executor makes it to this point in the code
path.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>