Remove GetSessionInfo Call
The GetSessionInfo call was added to the getSession
function which retrieves a session from the pool.
The call was used to validate the state of the session
before using it to ensure it was still healthy.
Unfortunately the call is exceedingly expensive with
preliminary tests showing nearly a second to perform the
single operation. With a transaction require requiring
several signings it is possible for it to take several
seconds for a transaction just to retrieve all of it signatures.

This change reverts the call and in future implementations
we can look into other ways of verifying if a session has
become stale, perhaps by analyzing failed transaction error
messages and ejecting the session on transaction failures.

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>