Discover CLI fails fast when given inappropriate certificate file
Currently, when a user incorrectly configures the discover CLI to use
an enrollment certificate that is not an x509 certificate, the CLI still
connects to the peer, and then authentication fails (obviously, since the peer cannot parse the certificate),
however the error returned to the client is still the obscure 'access denied'.

This PR makes the discover CLI to fail fast with an appropriate error when configured
with an enrollment certificate file that is not a certificate.

Change-Id: I255ec83ed4edc79957e0810b6146b07ec5957955
Signed-off-by: yacovm <yacovm@il.ibm.com>