Persist complete collection in the private blockstore (#2010)
- In the current code, we trim the collection to match the key-values
present in the booting snapshot. As per the current design
assumption, this suffices for transferring the private data to another
peer may bootstrap from the same or a newer snapshot at a later point
in time. However, a possibility of booting a peer from an older snapshot
(or from genesis block) cannot be ruled out and is good to be able to
support. This commit allows for the above by not removing the unused
key-values (if any) from the collection data for a transaction.

- Also fixes a bug by skipping a check; wherein we check for the
transaction validation flag before applying the missing private data.
Because, in the snapshot state, the data appears only from the
valid transactions

Signed-off-by: manish <manish.sethi@gmail.com>