FAB-18198 Ch.Part.API: Restructure registrar init
The Registrar now operates either with or without a system channel.
This creates two distinct (but slightly overlapping) initialization
flows. Here we restructure the Registrar initialization such that it
is clear which methods are common to both flows, and which methods
belong to the flow with or without the system channel.

For example, the current `initAppChannels` is split in two:
- initAppChannelsWithSystemChannel(), which assume the system channel exists, and
- initAppChannels(), which assumes the system channel does not exist.

This simplifies the code and makes it more readable, and tries to be
considerate of the fact that hopefully we'll be able to remove
the system channel related code (after the system channel functionality
is deprecated) so isolating it helps us later down the line.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I21b2e495745fec0859e2af998019b0a496e977e6