Copy Into Correct Slice To Ensure Keystore Password is Set
A previous PR refactored the clone command but
erronously overwrote the cloned value by copying
into the wrong slice before assigning the keystore
password as the empty slice. This caused future keystore
password lookups to be an empty value.

This change copies the keystore password value into the
correct slice.

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>