FAB-18283 Ch.Part.API: StatusReport on evicted etcdraft.Chain
When the channel participation API is used to list the details
for an application channel where the orderer has been removed
from the consenters set when the system channel exists, an
incorrect StatusReport of "member"/"active" is returned by the
API. It should return "config-tracker"/"inactive" at this point.

This commit fixes this problem by changing the StatusReport
return by an etcdraft.Chain after it is stopped and evicted
from the cluster.

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I28573859fc797489e721fd4ee8625de064be4235