Add function GetChannelConfig in cscc and use in peer client
CSCC has a function GetConfigBlock that is used by the peer client to
extract the orderer endpoints. A peer bootstrapped from a snapshot may
not have the config block until next channel config update is committed.

This commit introduces a function a cscc to return channel config and
makes use of this in the peer client.

Though, in general we do not encourage to add functions in these built-in
chanincodes, but here we cannot modify the existing function GetConfigBlock
as some external clients may be using this and it will break the backward
compatibility. However, the intention is to encourage the users to use
this new API to pave the way for the removal of the existing API.

Signed-off-by: manish <manish.sethi@gmail.com>