FAB-18291 Ch.Part.API: BlockPuller that can be stopped (#2050)
When a follower is started with a block that has bad endpoints,
it will cause the internal BlockPuller to retry a lot of times.
Calling Halt() on said follower will block until the BlockPuller
gives up, which can take a very long time. This causes an attempt
to Remove that channel via the participation API to block for too long.

This commit improves the BlockPuller by adding a channel to signal
it to stop retrying and exit. It also integrates with the
follower.Chain to make it more resposive to Halt().

Signed-off-by: Yoav Tock <tock@il.ibm.com>
Change-Id: I49354d94a28402ce69a469689c12ea595b3e0a7e