Optimize importing snapshot data into pvtdatastore
The pvtdata hashes in the snapshot files are maintained in the sort
order of key-hashes. However, the data in pvtdata store is primarily
maintained as group by block number. This causes inefficiencies during
importing the data in the pvtdata store primarily because without any
optimization, we end up performing write for each data item independently.

This would not have been as bad as such, if we write to leveldb as async
i/o, however, due to a bug previously observed in leveldb, we always
perform every write to leveldb as a sync write. This makes the overall
data loading into pvtdata store very poor performing because of
significantly high numbers of fsyncs.

In this commit, we adopt to a different approach where in we sort the
snapshot data by block number in a temporary space and group by desired
attributes for different entry types in memory and finally write the
entries in pvtdata store in batches.

Signed-off-by: manish <manish.sethi@gmail.com>