Add function TxIDExists and use in validation path
This commit adds a function in ledger TxIDExists and uses this in
validation path for checking duplicate txids.

The current validation code, in order to checks for a duplicate txid,
uses the API GetTransactionByID on ledger and relies on the error types
returned by this API. For a ledger that is bootstrapped from a snapshot,
this API needs to return an additional error if the txid belongs to a
block that is committed prior to the snapshot height.
Instead of exposing one more error type for validation code, we add an
explicit function for checking the existence of a TxID

Signed-off-by: manish <manish.sethi@gmail.com>