[FAB-18378] Log warning when peer is lagging behind and cannot catch up
Fabric 2.2 made gossip state transfer be disabled by default.

When a peer is a gossip follower, it does not pull blocks from the ordering service,
and relies on the leader peer of its organization to forward blocks directly or indirectly.

However, if a peer misses a block and the rest of the peers in its organization advance too much,
it will never catch up because its state transfer is disabled.

This commit makes the peer log a warning if it is lagging behind, and its configuration
might prevent it from obtaining blocks via state transfer and the orderer.

Change-Id: Ie7bc67cb13cb2d8e7ffab638a49e0365535865c2
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
(cherry picked from commit 5f61ff0d7a50ab0e9379b15051963ec7704d2954)