Fix test network tutorial to deploy go smart contract
In test-network, deployCC command requires the language and path
of the Chaincode. However, the tutorial does not specify anything
and the script fails to deploy Chaincode.
This PR fixes the command in the tutorial to deploy chaincode.

Signed-off-by: Yuki Kondo <yuki.kondo.ob@hitachi.com>
(cherry picked from commit 4fe9c764179a0c167cf2d56c286ecea936dadf51)

# Conflicts:
#	docs/source/test_network.md