Remove unreachable and unnecessary code in gossip membership
In gossip, alive messages are stored in membership stores, and they also have logical timestamps.

When a peer learns of an alive message with a higher logical timestamp of its latest alive message (for a specific peer),
it enters learnExistingMembers() and tentatively updates the latest envelope and gossip message of what is store in the membership store.

However, for completeness, the code checks if the membership store already contains an earlier instance, and if not, populates the membership store
with a new instance. This population is not needed, because the code only enters the function if it already knows the peer, hence the membership
store contains an earlier instance.

This commit simply removes the unneeded code path where the instance is unknown despite the fact that we know about the peer already.

Change-Id: Ia91cb35eabe18a8771ad12a30b06489892a2b49f
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
(cherry picked from commit a4a8520772cb2a1d2d73c6f53dbf8fae872c006f)